+++ src/main/java/capstone/safeat/member/dto/JwtMemberId.java
@@ -1,0 +1,5 @@
+package capstone.safeat.member.dto;
+
+public record JwtMemberId(Long memberId) {
+
+}
