+++ src/test/java/capstone/safeat/login/application/JwtProviderTest.java
@@ -1,7 +1,8 @@
-package capstone.safeat.member.application;
+package capstone.safeat.login.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import capstone.safeat.login.application.JwtProvider;
 import io.jsonwebtoken.Jwts.SIG;
 import io.jsonwebtoken.io.Encoders;
 import javax.crypto.SecretKey;
@@ -25,9 +26,9 @@ class JwtProviderTest {
   void 정상적으로_토큰을_생성한다() {
     final Long validId = 10L;
 
-    final String token = jwtProvider.createToken(validId);
+    final String token = jwtProvider.createAccessTokenWith(validId);
 
-    final Long actual = jwtProvider.parseToken(token);
+    final Long actual = jwtProvider.parseMemberId(token);
     assertThat(actual)
         .isEqualTo(validId);
   }
@@ -37,8 +38,8 @@ class JwtProviderTest {
     final Long validId = 10L;
     final JwtProvider differentKeyProvider = createRandomKeyProvider();
 
-    final String token = jwtProvider.createToken(validId);
+    final String token = jwtProvider.createAccessTokenWith(validId);
 
-    Assertions.assertThatThrownBy(() -> differentKeyProvider.parseToken(token));
+    Assertions.assertThatThrownBy(() -> differentKeyProvider.parseMemberId(token));
   }
 }
